requirements.txt
# FastAPI 웹 프레임워크
fastapi>=0.104.0
uvicorn>=0.24.0

# 데이터 검증 및 직렬화
pydantic>=2.5.0

# 환경 변수 관리
python-dotenv>=1.0.0

# LangChain 및 OpenAI
langchain>=0.0.350
langchain-openai>=0.0.2
langchain-community>=0.0.10
openai>=1.6.1

# 데이터베이스 연결
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=41.0.0 